# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - password
    type: object
  dto.ResetPwdRequest:
    properties:
      code:
        type: integer
      email:
        type: string
      new_password:
        type: string
    type: object
  dto.Response:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.ResponseToken:
    properties:
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: This is a coffee shop backend API
  title: Coffee Shop API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send OTP code to user email for password reset
      parameters:
      - description: User Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Request forgot password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Login Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Login new user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Register new user
      tags:
      - Authentication
  /auth/reset-password:
    patch:
      consumes:
      - application/json
      description: Reset user password using OTP code
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPwdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Reset password
      tags:
      - Auth
...
swagger: "2.0"